server.go + peer: use MessageConn interface instead of brontide.Conn
